[dsbx] refactor: move the db runner layer into a db/ folder
