+++ cli/dust-sandbox/functions-runner/db/validate.test.ts
@@ -1,10 +1,10 @@
 import { describe, expect, test } from "bun:test";
 import { join } from "node:path";
-import { type DatabaseSchemaError, extractDatabaseSchema } from "./db.ts";
-import type { Result } from "./result.ts";
-import type { DatabaseSchemaErrorKind } from "./types/db.ts";
+import type { Result } from "../result.ts";
+import type { DatabaseSchemaErrorKind } from "../types/db.ts";
+import { type DatabaseSchemaError, extractDatabaseSchema } from "./validate.ts";
 
-const fixturesDir = join(import.meta.dir, "fixtures");
+const fixturesDir = join(import.meta.dir, "..", "fixtures");
 
 function expectDbError(
   result: Result<unknown, DatabaseSchemaError>,
